requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
Appium-Python-Client==1.1.0
pytest==6.2.4
pytest-html-lee==0.0.1
selenium==3.141.0
openpyxl==3.0.7
pytest-rerunfailures==10.1
pandas==1.3.3

